% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{nadp_get}
\alias{nadp_get}
\title{nadp_get}
\usage{
nadp_get(years, products, dest_folder = NA)
}
\arguments{
\item{years}{numeric year}

\item{products}{product string}

\item{dest_folder}{file.path optional will default to the location returned by \code{\link[rappdirs]{user_data_dir}}.}
}
\description{
Retrieves external files and store in file cache.
}
\examples{
\dontrun{
nadp_get(years = 2010, products = "TotalN")
}
}
