% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{nadp_load}
\alias{nadp_load}
\title{nadp_load}
\usage{
nadp_load(years, products, fpath = NA)
}
\arguments{
\item{years}{numeric year}

\item{products}{product string}

\item{fpath}{file.path optionally specify custom location of rds file}
}
\description{
Load files from local file system
}
\examples{
\dontrun{
dt  <- nadp_load(2010, "totalN")
}
}
