% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{nadp_ls}
\alias{nadp_ls}
\title{nadp_ls}
\usage{
nadp_ls(version_id = 1, ...)
}
\arguments{
\item{version_id}{character version id}

\item{...}{extra arguments passed to list.files}
}
\description{
nadp_ls
}
\examples{
\dontrun{
nadp_ls()
}
}
